requirements.txt
Django==1.5.1
PyChef==0.2.1
django-breadcrumbs==1.1.3-p1
django-redis==3.2
redis==2.7.3
wsgiref==0.1.2
